Create common test utils, change AccountTransaction::Declare construction
