+++ crates/blockifier/src/transaction.rs
@@ -2,6 +2,7 @@ pub mod account_transaction;
 pub mod constants;
 pub mod errors;
 pub mod objects;
+pub mod test_utils;
 pub mod transaction_execution;
 pub mod transaction_types;
 pub mod transaction_utils;
